Separating storage into smaller chunks
