+++ contracts/mixnet/src/gateways/mod.rs
@@ -1,2 +1,3 @@
 pub mod queries;
+pub mod storage;
 pub mod transactions;
