+++ contracts/mixnet/src/mixnodes/mod.rs
@@ -3,3 +3,4 @@ pub mod bonding_transactions;
 pub mod delegation_queries;
 pub mod delegation_transactions;
 pub mod layer_queries;
+pub mod storage;
